Set object values, not just ids, for related objects.
If we only set the id values, the database will be properly updated when
the data is saved, but, until it is refreshed from the database, those
objects will be null, which is not what we want.  We want to use those
objects to easily get information about the related objects.